Add tests for INSTRONA
